backend/python/social-network/src/requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-configurations==2.1
django-extensions==2.1.6
django-filter==2.1.0
django-imagekit==4.0.2
django-nose==1.4.6
django-phonenumber-field==2.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.5
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nose==1.3.7
oauthlib==3.0.1
openapi-codec==1.3.2
phonenumbers==8.10.10
pilkit==2.0
Pillow==6.0.0
psycopg2==2.8.2
ptvsd==4.2.10
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.25.2
